improve signature services
