+++ xrpl4j-crypto-parent/xrpl4j-crypto-core/src/main/java/org/xrpl/xrpl4j/crypto/core/signing/AbstractDelegatedSignatureService.java
@@ -23,98 +23,37 @@ public abstract class AbstractDelegatedSignatureService implements DelegatedSign
 
   protected final Logger logger = LoggerFactory.getLogger(this.getClass());
 
-  private final KeyStoreType keyStoreType;
+  private SignatureUtils signatureUtils = null;
 
-  private final SignatureUtils signatureUtils;
+  private AddressUtils addressService = null;
 
-  private final AddressUtils addressService;
+  private final DelegatedTransactionSigner delegatedTransactionSigner;
+  private final DelegatedTransactionVerifier delegatedTransactionVerifier;
 
-  /**
-   * Required-args Constructor.
-   *
-   * @param keyStoreType   The {@link KeyStoreType} for this service.
-   * @param signatureUtils An {@link SignatureUtils} for help with signing.
-   * @param addressService A {@link AddressUtils}.
-   */
-  public AbstractDelegatedSignatureService(
-    final KeyStoreType keyStoreType,
-    final SignatureUtils signatureUtils,
-    final AddressUtils addressService
-  ) {
-    this.keyStoreType = Objects.requireNonNull(keyStoreType);
-    this.signatureUtils = Objects.requireNonNull(signatureUtils);
-    this.addressService = Objects.requireNonNull(addressService);
+  public AbstractDelegatedSignatureService(DelegatedTransactionSigner delegatedTransactionSigner, DelegatedTransactionVerifier delegatedTransactionVerifier) {
+    this.delegatedTransactionSigner = delegatedTransactionSigner;
+    this.delegatedTransactionVerifier = delegatedTransactionVerifier;
   }
 
+
   @Override
-  public final KeyStoreType keyStoreType() {
-    return this.keyStoreType;
+  public PublicKey getPublicKey(KeyMetadata keyMetadata) {
+    return null;
   }
 
   @Override
   public <T extends Transaction> SingleSingedTransaction<T> sign(KeyMetadata keyMetadata, T transaction) {
-    Objects.requireNonNull(keyMetadata);
-    Objects.requireNonNull(transaction);
-
-    final UnsignedByteArray signableTransactionBytes = this.signatureUtils.toSignableBytes(transaction);
-    final Signature signature = this.signingHelper(keyMetadata, signableTransactionBytes).transactionSignature();
-    return this.signatureUtils.addSignatureToTransaction(transaction, signature);
+    return delegatedTransactionSigner.sign(keyMetadata, transaction);
   }
 
   @Override
-  public Signature sign(final KeyMetadata keyMetadata, final UnsignedClaim unsignedClaim) {
-    Objects.requireNonNull(keyMetadata);
-    Objects.requireNonNull(unsignedClaim);
-
-    final UnsignedByteArray signableBytes = signatureUtils.toSignableBytes(unsignedClaim);
-    return this.signingHelper(keyMetadata, signableBytes).transactionSignature();
+  public Signature sign(KeyMetadata keyMetadata, UnsignedClaim unsignedClaim) {
+    return null;
   }
 
   @Override
   public <T extends Transaction> SignatureWithKeyMetadata multiSign(KeyMetadata keyMetadata, T transaction) {
-    Objects.requireNonNull(keyMetadata);
-    Objects.requireNonNull(transaction);
-
-    final Address address = addressService.deriveAddress(this.getPublicKey(keyMetadata));
-    final UnsignedByteArray signableTransactionBytes = this.signatureUtils.toMultiSignableBytes(transaction, address);
-
-    return this.signingHelper(keyMetadata, signableTransactionBytes);
-  }
-
-  /**
-   * Helper to generate a signature based upon an {@link UnsignedByteArray} of transaction bytes.
-   *
-   * @param keyMetadata              A {@link KeyMetadata} for the signing key.
-   * @param signableTransactionBytes A {@link UnsignedByteArray} of transaction bytes.
-   *
-   * @return A {@link SignatureWithPublicKey}.
-   */
-  private SignatureWithKeyMetadata signingHelper(
-    final KeyMetadata keyMetadata, final UnsignedByteArray signableTransactionBytes
-  ) {
-    Objects.requireNonNull(keyMetadata);
-    Objects.requireNonNull(signableTransactionBytes);
-
-    final PublicKey publicKey = this.getPublicKey(keyMetadata);
-    final Signature signature;
-    switch (publicKey.versionType()) {
-      case ED25519: {
-        signature = this.edDsaSign(keyMetadata, signableTransactionBytes);
-        break;
-      }
-      case SECP256K1: {
-        signature = this.ecDsaSign(keyMetadata, signableTransactionBytes);
-        break;
-      }
-      default: {
-        throw new IllegalArgumentException("Unhandled PrivateKey VersionType: {}" + keyMetadata);
-      }
-    }
-
-    return SignatureWithKeyMetadata.builder()
-      .signingKeyMetadata(keyMetadata)
-      .transactionSignature(signature)
-      .build();
+    return null;
   }
 
   @Override
@@ -240,12 +179,14 @@ public abstract class AbstractDelegatedSignatureService implements DelegatedSign
    * @param signableTransactionBytes A {@link UnsignedByteArray} to sign.
    *
    * @return A {@link Signature} with data that can be used to submit a transaction to the XRP Ledger.
-   */
-  protected abstract Signature edDsaSign(
+   *//*
+  protected Signature edDsaSign(
     KeyMetadata privateKeyMetadata, UnsignedByteArray signableTransactionBytes
-  );
+  ) {
+    return delegatedTransactionSigner.
+  }
 
-  /**
+  *//**
    * Does the actual work of computing a signature using a secp256k1 private-key, as locatable using {@code
    * privateKeyMetadata}.
    *
@@ -253,10 +194,10 @@ public abstract class AbstractDelegatedSignatureService implements DelegatedSign
    * @param signableTransactionBytes A {@link UnsignedByteArray} to sign.
    *
    * @return A {@link Signature} with data that can be used to submit a transaction to the XRP Ledger.
-   */
+   *//*
   protected abstract Signature ecDsaSign(
     KeyMetadata privateKeyMetadata, UnsignedByteArray signableTransactionBytes
-  );
+  );*/
 
   /**
    * Verify a signature.
