+++ xrpl4j-crypto-parent/xrpl4j-crypto-core/src/main/java/org/xrpl/xrpl4j/crypto/core/signing/AbstractDelegatedTransactionVerifier.java
@@ -1,0 +1,4 @@
+package org.xrpl.xrpl4j.crypto.core.signing;
+
+public abstract class AbstractDelegatedTransactionVerifier implements DelegatedTransactionVerifier {
+}
