+++ xrpl4j-crypto-parent/xrpl4j-crypto-core/src/main/java/org/xrpl/xrpl4j/crypto/core/signing/DelegatedTransactionSigner.java
@@ -1,6 +1,7 @@
 package org.xrpl.xrpl4j.crypto.core.signing;
 
 import org.xrpl.xrpl4j.crypto.core.KeyMetadata;
+import org.xrpl.xrpl4j.crypto.core.keys.PublicKey;
 import org.xrpl.xrpl4j.model.client.channels.UnsignedClaim;
 import org.xrpl.xrpl4j.model.transactions.Transaction;
 
@@ -11,6 +12,17 @@ import org.xrpl.xrpl4j.model.transactions.Transaction;
  * manner (e.g., by deriving it from a secret value).
  */
 public interface DelegatedTransactionSigner {
+
+  /**
+   * Accessor for the public-key corresponding to the supplied key meta-data. This method exists to support
+   * implementations that hold private-key material internally, yet need a way for external callers to determine the
+   * actual public key for signature verification or other purposes.
+   *
+   * @param keyMetadata A {@link KeyMetadata} for a key-pair.
+   *
+   * @return A {@link PublicKey}.
+   */
+  PublicKey getPublicKey(KeyMetadata keyMetadata);
 
   /**
    * Obtain a signature for the supplied transaction using the private-key that corresponds to {@code keyMetadata}.
